refactor external logging to be a microservice
