+++ services/external_logging/supervisor.conf
@@ -1,0 +1,7 @@
+[program:cron]
+command=/usr/bin/env python baselayer/services/cron/cron.py %(ENV_FLAGS)s
+environment=PYTHONPATH=".",PYTHONUNBUFFERED="1"
+startretries=0
+startsecs=1
+stdout_logfile=log/cron.log
+redirect_stderr=true
